bgbot/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
authlib==1.3.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
